+++ source/plugins/first.run/index.js
@@ -5,7 +5,7 @@ let firstrun = {
 
 	isFirstRun()
 	{
-		return studio.settings.loadValue('firstrun', 'firstrun', true);
+		return studio.settings.loadValue('firstrun', 'firstRun', true);
 	},
 
 	showFirstRun()
@@ -13,7 +13,7 @@ let firstrun = {
 		studio.workspace.showDialog(FirstRunDialog,{width:550});
 		studio.settings.storeValue('firstrun', 'firstRun', false);
 	}
-}
+};
 
 export function setup(options, imports, register)
 {